[iOS Sim] Layout Test scrollingcoordinator/ios/ui-scroll-fixed.html is a flaky failure https://bugs.webkit.org/show_bug.cgi?id=196612 rdar://problem/49612867
Reviewed by Antti Koivisto.

This test hit a really obscure bug where a combination of an immediate scroll, and
an ancestor reposition left the layer position of a position:fixed layer unchanged.
The position of this layer in the UI process had been previously modified by
the scrolling tree for the scroll, but because the WebContent-side mutations left
the actual position unchanged, we'd never apply a new position via a commit, so left
the layer in the wrong location.

Removing the m_wasScrolledByDelegatedScrollingSincePreviousCommit check in ScrollingTree::applyLayerPositionsAfterCommit()
fixes this, but has perf implications. Alternative fixes require complex state tracking.
Since this is so hard to hit with noisy user scrolling, just change the test to avoid
the perfect storm of scrolls and offsets.

* platform/ios-simulator-wk2/TestExpectations:
* scrollingcoordinator/ios/ui-scroll-fixed-expected.html:
* scrollingcoordinator/ios/ui-scroll-fixed.html:


Identifier: 216517@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@251251 268f45cc-cd09-0410-ab3c-d52691b4dbfc