Post increment/decrement should only call ToNumber once https://bugs.webkit.org/show_bug.cgi?id=202711
Reviewed by Saam Barati.

JSTests:

* stress/postinc-custom-valueOf.js: Added.
(postInc):
(postDec):

Source/JavaScriptCore:

The problem is that we first called ToNumber on the object being incremented (to have the result that we'll eventually return), but we then do emitIncOrDec on the original object, which can call ToNumber again.
Instead we must do the ToNumber once, then copy its result, emitIncOrDec on the copy, put the copy back in the original location, and finally return the old value.
Since the result of ToNumber is guaranteed not to be an object, emitIncOrDec won't call ToNumber a second time.

* bytecompiler/NodesCodegen.cpp:
(JSC::emitPostIncOrDec):


Identifier: 216609@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@251371 268f45cc-cd09-0410-ab3c-d52691b4dbfc