JSON.parse has bad is array assert https://bugs.webkit.org/show_bug.cgi?id=203207 <rdar://problem/56366913>
Reviewed by Yusuke Suzuki.

JSTests:

* stress/json-parse-array-prototype-is-array-assert.js: Added.
(assert):

Source/JavaScriptCore:

In r250860, we updated JSON.parse to be more spec compliant with how it
handles arrays. However, we also updated an assertion in an improper way,
where our assert was not accounting for the check we performed to take
that control flow path.

* runtime/JSONObject.cpp:
(JSC::Walker::walk):


Identifier: 216631@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@251394 268f45cc-cd09-0410-ab3c-d52691b4dbfc