WebSWServerToContextConnection should not assert when failing loads at destruction time https://bugs.webkit.org/show_bug.cgi?id=203243
Reviewed by Alex Christensen.

On WebSWServerToContextConnection destruction, we move the fetch task map and fail the tasks.
At destruction of the tasks, which happens synchronously, they will try to unregister themselves
and the assertion that the task is in the map will fail.
To fix that, add a specific contextClosed method that will clear the task connection weak pointer.

* NetworkProcess/ServiceWorker/ServiceWorkerFetchTask.cpp:
(WebKit::ServiceWorkerFetchTask::contextClosed):
* NetworkProcess/ServiceWorker/ServiceWorkerFetchTask.h:
* NetworkProcess/ServiceWorker/WebSWServerToContextConnection.cpp:
(WebKit::WebSWServerToContextConnection::~WebSWServerToContextConnection):


Identifier: 216671@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@251434 268f45cc-cd09-0410-ab3c-d52691b4dbfc